refactor: move Dense layers to dense.py
Layers SpectralDense and FrobeniusDense are moved to the new file dense.py
